infrastructure/k8s/monitoring/requirements.txt
# Requirements for Rollback Automation Service
kubernetes==28.1.0
prometheus-api-client==0.5.3
PyYAML==6.0.1
requests==2.31.0
flask==3.0.0
prometheus-client==0.19.0